refactoring(15): extract renderRepro from ResultsPage (medium-risk seam)
Move the reproducibility section (rich ORS chain / legacy panel + repro-log
toolbar/panel) into resultSections.tsx as renderRepro({...}). Body verbatim
from the container — the closed-over repro-log state, its setter/loader,
summary, selectedId and t become destructured params; the internal
handleToggle/Refresh handlers move unchanged. renderOrsChain/renderLegacyRepro
now resolve in-file (dropped from the container import). ResultsPage 1446->1327.
