+++ ari-core/ari/viz/frontend/src/components/Results/ResultsPage.tsx
@@ -28,7 +28,7 @@ import { Card } from '../common/Card';
 import { Button } from '../common/Button';
 import { Badge } from '../common/Badge';
 import { PublishYamlEditor } from './PublishYamlEditor';
-import { renderOrsChain, renderLegacyRepro, renderContext, renderFigures, renderReviewScores } from './resultSections';
+import { renderContext, renderFigures, renderReviewScores, renderRepro } from './resultSections';
 
 
 export function ResultsPage() {
@@ -825,135 +825,6 @@ export function ResultsPage() {
   //   - "legacy": when only ``reproducibility_report`` (the legacy
   //     pre-§4.1 format) is present. Falls back to the previous flat
   //     verdict + key-value layout.
-  const renderRepro = () => {
-    if (!summary) return null;
-
-    const repro = summary.reproducibility_report || summary.repro;
-    const orsRubric = (summary as any).ors_rubric as Record<string, any> | undefined;
-    const orsGrade = (summary as any).ors_grade as Record<string, any> | undefined;
-    const orsPhase1 = (summary as any).ors_phase1 as Record<string, any> | undefined;
-    const orsReplicator = (summary as any).ors_replicator as Record<string, any> | undefined;
-    const orsSeed = (summary as any).ors_seed as Record<string, any> | undefined;
-    const orsRubricMeta = (summary as any).ors_rubric_meta as Record<string, any> | undefined;
-    const richMode = !!(orsRubric || orsGrade || orsPhase1 || orsReplicator || orsSeed || orsRubricMeta);
-
-    const handleToggleReproLog = () => {
-      if (reproLogOpen) { setReproLogOpen(false); return; }
-      setReproLogOpen(true);
-      if (selectedId) loadReproLog(selectedId);
-    };
-    const handleRefreshReproLog = () => {
-      if (selectedId) loadReproLog(selectedId);
-    };
-
-    const reproLogPanel = reproLogOpen && (
-      <div
-        style={{
-          border: '1px solid var(--border)',
-          borderRadius: 4,
-          marginBottom: 8,
-          maxHeight: 320,
-          overflow: 'auto',
-          background: 'var(--bg)',
-        }}
-      >
-        {reproLogPath && (
-          <div
-            style={{
-              fontSize: '.7rem',
-              color: 'var(--muted)',
-              padding: '4px 8px',
-              borderBottom: '1px solid var(--border)',
-              fontFamily: 'monospace',
-            }}
-          >
-            {reproLogPath}
-          </div>
-        )}
-        {reproLogLoading ? (
-          <div style={{ padding: 8, fontSize: '.8rem', color: 'var(--muted)' }}>
-            {t('repro_log_loading')}
-          </div>
-        ) : reproLogContent ? (
-          <pre
-            style={{
-              margin: 0, padding: '6px 10px',
-              fontSize: '.72rem', lineHeight: 1.45,
-              fontFamily: 'monospace',
-              whiteSpace: 'pre-wrap', wordBreak: 'break-all',
-              color: 'var(--text)',
-            }}
-          >
-            {reproLogContent}
-          </pre>
-        ) : (
-          <div style={{ padding: 8, fontSize: '.8rem', color: 'var(--muted)' }}>
-            {t('repro_log_empty')}
-          </div>
-        )}
-      </div>
-    );
-
-    const toolbar = (
-      <div style={{ display: 'flex', gap: 6, marginBottom: 8 }}>
-        <Button
-          onClick={handleToggleReproLog}
-          style={{ fontSize: '.75rem', padding: '3px 10px' }}
-        >
-          {reproLogOpen ? t('repro_log_hide') : t('repro_log_show')}
-        </Button>
-        {reproLogOpen && (
-          <Button
-            onClick={handleRefreshReproLog}
-            style={{ fontSize: '.75rem', padding: '3px 8px' }}
-            disabled={reproLogLoading}
-            title={t('repro_log_refresh')}
-          >
-            {t('repro_log_refresh')}
-          </Button>
-        )}
-      </div>
-    );
-
-    if (richMode) {
-      return (
-        <Card style={{ marginBottom: 16 }}>
-          <div className="card-title">{t('verify_title')}</div>
-          {renderOrsChain({
-            repro,
-            orsRubric, orsGrade, orsPhase1, orsReplicator, orsSeed, orsRubricMeta,
-            ckptId: selectedId || '',
-            reproLog: {
-              open: reproLogOpen,
-              loading: reproLogLoading,
-              content: reproLogContent,
-              path: reproLogPath,
-            },
-            onToggleLog: handleToggleReproLog,
-            onRefreshLog: handleRefreshReproLog,
-            t,
-          })}
-        </Card>
-      );
-    }
-
-    // ── Legacy renderer (pre-§4.1 reproducibility_report shape) ─────────
-    return (
-      <Card style={{ marginBottom: 16 }}>
-        <div className="card-title">{t('verify_title')}</div>
-        {toolbar}
-        {reproLogPanel}
-        {repro ? (
-          renderLegacyRepro({ repro, t })
-        ) : (
-          <div style={{ color: 'var(--muted)', fontSize: '.85rem' }}>
-            {t('no_repro')}
-          </div>
-        )}
-      </Card>
-    );
-  };
-
   // Render experiment context
   // Render Experiment Artifact Repository (EAR) section — issue #4
   const renderEAR = () => {
@@ -1414,7 +1285,17 @@ export function ResultsPage() {
           <>
             {renderPaper()}
             {renderReviewScores({ summary, t })}
-            {renderRepro()}
+            {renderRepro({
+              summary,
+              selectedId,
+              reproLogOpen,
+              reproLogContent,
+              reproLogPath,
+              reproLogLoading,
+              setReproLogOpen,
+              loadReproLog,
+              t,
+            })}
             {renderEAR()}
             {renderContext({ summary, t })}
             {renderFigures({ summary })}
